requirements.txt
Flask==1.0.2
python-telegram-bot==10.1.0
requests==2.18.4
zappa==0.45.1
